Move part of api.hpp to api_converters.hpp
